+++ wpf/Learn.MultipleFrameworks/Learn.MultipleFrameworks/ViewModels/NumericKeyboardViewModel.cs
@@ -13,7 +13,7 @@ public class NumericKeyboardViewModel : KeyboardViewModel
             return DefaultValue;
         }
 
-        if (int.TryParse(input, out _))
+        if (long.TryParse(input, out _))
         {
             return InputUtilities.TrimStartZero(input);
         }
